Remove tile threshold function. Fix vitests.
