+++ app/app.module.ts
@@ -3,7 +3,6 @@ import {BrowserModule} from '@angular/platform-browser';
 import {AppRoutingModule} from './app.routing.module';
 import {AppComponent} from './app.component';
 import {provideHttpClient} from "@angular/common/http";
-import {SpeedDialModule} from "primeng/speeddial";
 import {DialogModule} from "primeng/dialog";
 import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
 import {AnimateOnScroll} from "primeng/animateonscroll";
@@ -249,7 +248,6 @@ export const initializeServices = () => {
         AppConfirmPopupComponent,
         SearchCompletionPopupComponent,
         SimfilExpressionInputComponent,
-        SpeedDialModule,
         DialogModule,
         DynamicDialogModule,
         FormsModule,
