+++ app/inspection/inspection.tree.component.ts
@@ -24,6 +24,7 @@ import {JumpTargetService} from "../search/jump.service";
 import {stripFeatureInspectionTarget} from "../shared/tile-feature-id";
 import {FeatureSearchService} from "../search/feature.search.service";
 import type {FeatureSearchMapLayerRef} from "../shared/feature-search-state";
+import {inspectionSearchNumberLiteral} from "./inspection-search.util";
 
 /** Column definition used by the inspection tree's generic table renderer. */
 export interface Column {
@@ -39,24 +40,6 @@ export class FeatureFilterOptions {
     filterByValues: boolean = true;
     filterOnlyFeatureIds: boolean = false;
     filterGeometryEntries: boolean = false;
-}
-
-/** Converts inspection numeric values to Simfil numeric literals without quoting integer BigInts. */
-export function inspectionSearchNumberLiteral(value: unknown): string | undefined {
-    if (typeof value === "bigint") {
-        return value.toString();
-    }
-    if (typeof value === "number" && Number.isFinite(value)) {
-        return String(value);
-    }
-    if (typeof value !== "string") {
-        return undefined;
-    }
-    const trimmed = value.trim();
-    return /^[-+]?(?:\d+(?:\.\d*)?|\.\d+)(?:[eE][-+]?\d+)?$/.test(trimmed) &&
-        Number.isFinite(Number(trimmed))
-        ? trimmed
-        : undefined;
 }
 
 @Component({
