+++ app/mapview/low-fi-tile-thresholds.ts
